+++ domain/src/main/java/org/plavelo/prs/domain/ChannelId.kt
@@ -1,3 +1,3 @@
 package org.plavelo.prs.domain
 
-data class ChannelId(val value: Int)
+data class ChannelId(val value: Long)
